DAH-2293 test all in one file
